chore: change structure
